requirements.txt
asgiref==3.3.1
click==7.1.2
DAWG-Python==0.7.2
Django==3.1.7
docopt==0.6.2
gunicorn==20.0.4
joblib==1.0.1
nltk==3.5
numpy==1.20.1
pandas==1.2.2
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dateutil==2.8.1
pytz==2021.1
regex==2020.11.13
schedule==1.0.0
scikit-learn==0.24.1
scipy==1.6.1
selenium==3.141.0
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tqdm==4.57.0
urllib3==1.26.3
